refactor(roadmap): update roadmap imports to use schemas package and remove deprecated references
