+++ packages/schemas/roadmap.ts
@@ -1,0 +1,16 @@
+export interface Roadmap {
+  id: string;
+  title: string;
+  description: string;
+  status: string;
+  labels: string[];
+  author: string;
+  createdAt: string;
+  updatedAt: string;
+  url: string;
+}
+
+export interface RoadmapListResponse {
+  success: boolean;
+  roadmap: Roadmap[];
+}
