+++ workers/roadmap/package.json
@@ -10,6 +10,7 @@
     "cf-typegen": "wrangler types"
   },
   "dependencies": {
+    "@repo/schemas": "workspace:*",
     "chanfana": "^2.6.3",
     "hono": "^4.9.9",
     "octokit": "^5.0.3",
